authorization: Fix endpoint path
The path for the endpoint shown in the example is slightly confusing
as it is a traditional REST endpoint. Usually MCP servers are free
to define _one_ MCP endpoint. Commonly `/`, `/sse`, `/messages` or `/mcp`
are used. In this ase we choose `/mcp` to stay away from any indication
if we are using SSE or not.