Switch to "mint" theme
The "willow" theme places the logo, version select box, and light/dark
switch all side by side in a constrained space at the top of the
sidebar, which causes rendering issues for each.

In contrast, the "mint" theme places each in a page-wide top bar,
eliminating space issues.  The "mint" theme also places the search box
in that top bar, allowing more of the sidebar to appear above the fold.