docs: Fix h5, h6 style
I noticed that mintlify does not correctly create id="" tags for h5 and
h6 headers. This means that we do not correctly trigger our styles on these
as they depend on the presence of id. For now we fix this by including h5 for
non-id, while we follow up with mintlify to fix this.