Further clarify SEP process (#987)
* Fix typos and formatting

* Reorient section headings

This swaps the "SEP Workflow" and "Submitting a SEP" headings so that
"Submitting a SEP" is the umbrella for "SEP Workflow", "SEP States",
etc.

* Bump up "SEP Format" section

This commit renames the "What belongs in a successful SEP?" heading to
"SEP Format" and places that section just below the "SEP Workflow"
section, which alludes to format.  This commit also removes the
redundant (and less thorough) description of the SEP format from the
"SEP Workflow" section.

* Refine "SEP Workflow" steps

This commit rewords the "SEP Workflow" section for clarity.  This commit
does not change the actual information content.

* Refine "SEP Format" items

This commit rewords the "SEP Format" section for clarity.  This commit
does not change the information content; however, it does swap the order
of the "Rationale" and "Specification" points, placing "Specification"
before "Rationale" because "Rationale" explains the design decisions of
"Specification".