Tweak layout of ecosystem stats
This commit primarily changes the layout of the ecosystem stats such
that on small screens the stats are displayed as a column of three
cards, and on larger screens that stats are displayed as a 2x2 grid of
cards with the "servers" card occupying two columns.

This commit also changes some of the styling to better suit the layout.

Co-authored-by: Cesar Garcia <Cesarponce19544@gmail.com>