Fix CTA button color on landing page
This commit fixes white-on-white text in the landing page's CTA button
when using light mode.

The background color of the button was removed in
cd7adf01543f50acfb03feae2c4aca1b25abf1a2, but the white-on-white text
was prevented due to a typo (a missing semi-colon).  The typo was
subsequently fixed in a1d6bed769d70f5b0167d8ae9ac994c768f745ac.